Removed uint64_t in favor of size_t for indices. Also removed base matrix class to make things easier

+++ include/sabai/metrics.hpp
@@ -8,8 +8,7 @@ template <typename T> double norm(const DynamicVector<T> &vector) {
   return sqrt(dot(vector, vector));
 }
 
-template <typename T, uint64_t N>
-double norm(const StaticVector<T, N> &vector) {
+template <typename T, size_t N> double norm(const StaticVector<T, N> &vector) {
   return sqrt(static_cast<double>(dot(vector, vector)));
 }
 } // namespace sabai
